+++ banyand/trace/block_writer.go
@@ -135,25 +135,31 @@ func (sw *writers) getWriters(tagName string) (*writer, *writer) {
 }
 
 type blockWriter struct {
-	writers                        writers
-	traceIDs                       []string
+	traceIDFilter                  *traceIDFilter
 	tagType                        tagType
+	writers                        writers
+	tidLast                        string
+	tidFirst                       string
 	metaData                       []byte
 	primaryBlockData               []byte
 	primaryBlockMetadata           primaryBlockMetadata
-	totalBlocksCount               uint64
-	maxTimestamp                   int64
 	totalUncompressedSpanSizeBytes uint64
 	totalCount                     uint64
 	minTimestamp                   int64
 	totalMinTimestamp              int64
 	totalMaxTimestamp              int64
 	minTimestampLast               int64
+	maxTimestamp                   int64
+	totalBlocksCount               uint64
+	hasWrittenBlocks               bool
 }
 
 func (bw *blockWriter) reset() {
 	bw.writers.reset()
-	bw.traceIDs = bw.traceIDs[:0]
+	if bw.traceIDFilter != nil {
+		bw.traceIDFilter.reset()
+		bw.traceIDFilter = nil
+	}
 	if bw.tagType == nil {
 		bw.tagType = make(tagType)
 	} else {
@@ -170,17 +176,27 @@ func (bw *blockWriter) reset() {
 	bw.primaryBlockData = bw.primaryBlockData[:0]
 	bw.metaData = bw.metaData[:0]
 	bw.primaryBlockMetadata.reset()
+	bw.tidFirst = ""
+	bw.tidLast = ""
+	bw.hasWrittenBlocks = false
 }
 
-func (bw *blockWriter) MustInitForMemPart(mp *memPart) {
+func (bw *blockWriter) MustInitForMemPart(mp *memPart, traceSize int) {
 	bw.reset()
 	bw.writers.mustCreateTagWriters = mp.mustCreateMemTagWriters
 	bw.writers.metaWriter.init(&mp.meta)
 	bw.writers.primaryWriter.init(&mp.primary)
 	bw.writers.spanWriter.init(&mp.spans)
+	if traceSize > 0 {
+		bw.traceIDFilter = &traceIDFilter{
+			filter: generateTraceIDBloomFilter(),
+		}
+		bw.traceIDFilter.filter.SetN(traceSize)
+		bw.traceIDFilter.filter.ResizeBits((traceSize*filter.B + 63) / 64)
+	}
 }
 
-func (bw *blockWriter) mustInitForFilePart(fileSystem fs.FileSystem, path string, shouldCache bool) {
+func (bw *blockWriter) mustInitForFilePart(fileSystem fs.FileSystem, path string, shouldCache bool, traceSize int) {
 	bw.reset()
 	fileSystem.MkdirPanicIfExist(path, storage.DirPerm)
 	bw.writers.mustCreateTagWriters = func(name string) (fs.Writer, fs.Writer) {
@@ -193,6 +209,13 @@ func (bw *blockWriter) mustInitForFilePart(fileSystem fs.FileSystem, path string
 	bw.writers.metaWriter.init(fs.MustCreateFile(fileSystem, filepath.Join(path, metaFilename), storage.FilePerm, shouldCache))
 	bw.writers.primaryWriter.init(fs.MustCreateFile(fileSystem, filepath.Join(path, primaryFilename), storage.FilePerm, shouldCache))
 	bw.writers.spanWriter.init(fs.MustCreateFile(fileSystem, filepath.Join(path, spansFilename), storage.FilePerm, shouldCache))
+	if traceSize > 0 {
+		bw.traceIDFilter = &traceIDFilter{
+			filter: generateTraceIDBloomFilter(),
+		}
+		bw.traceIDFilter.filter.SetN(traceSize)
+		bw.traceIDFilter.filter.ResizeBits((traceSize*filter.B + 63) / 64)
+	}
 }
 
 func (bw *blockWriter) MustWriteTrace(tid string, spans [][]byte, tags [][]*tagValue, timestamps []int64, spanIDs []string) {
@@ -210,21 +233,22 @@ func (bw *blockWriter) mustWriteBlock(tid string, b *block) {
 	if b.Len() == 0 {
 		return
 	}
-	hasWrittenBlocks := len(bw.traceIDs) > 0
-	var tidLast string
-	var isSeenTid bool
-
-	if hasWrittenBlocks {
-		tidLast = bw.traceIDs[len(bw.traceIDs)-1]
-		if tid < tidLast {
-			logger.Panicf("the tid=%s cannot be smaller than the previously written tid=%s", tid, tidLast)
-		}
-		isSeenTid = tid == tidLast
+	if tid < bw.tidLast {
+		logger.Panicf("the tid=%s cannot be smaller than the previously written tid=%s", tid, bw.tidLast)
 	}
+	hasWrittenBlocks := bw.hasWrittenBlocks
+	if !hasWrittenBlocks {
+		bw.tidFirst = tid
+		bw.hasWrittenBlocks = true
+	}
+	isSeenTid := tid == bw.tidLast
+	bw.tidLast = tid
 
 	bm := generateBlockMetadata()
 	b.mustWriteTo(tid, bm, &bw.writers)
-	bw.traceIDs = append(bw.traceIDs, tid)
+	if bw.traceIDFilter != nil && bw.traceIDFilter.filter != nil {
+		bw.traceIDFilter.filter.Add(convert.StringToBytes(tid))
+	}
 	bw.tagType.copyFrom(bm.tagType)
 	tm := &bm.timestamps
 	if bw.totalCount == 0 || tm.min < bw.totalMinTimestamp {
@@ -258,11 +282,13 @@ func (bw *blockWriter) mustWriteBlock(tid string, b *block) {
 
 func (bw *blockWriter) mustFlushPrimaryBlock(data []byte) {
 	if len(data) > 0 {
-		bw.primaryBlockMetadata.mustWriteBlock(data, bw.traceIDs[0], &bw.writers)
+		bw.primaryBlockMetadata.mustWriteBlock(data, bw.tidFirst, &bw.writers)
 		bw.metaData = bw.primaryBlockMetadata.marshal(bw.metaData)
 	}
+	bw.hasWrittenBlocks = false
 	bw.minTimestamp = 0
 	bw.maxTimestamp = 0
+	bw.tidFirst = ""
 }
 
 func (bw *blockWriter) mustWriteRawBlock(r *rawBlock) {
@@ -270,18 +296,19 @@ func (bw *blockWriter) mustWriteRawBlock(r *rawBlock) {
 	if bm.count == 0 {
 		return
 	}
-	hasWrittenBlocks := len(bw.traceIDs) > 0
-	var tidLast string
-	var isSeenTid bool
-
-	if hasWrittenBlocks {
-		tidLast = bw.traceIDs[len(bw.traceIDs)-1]
-		if bm.traceID < tidLast {
-			logger.Panicf("the tid=%s cannot be smaller than the previously written tid=%s", bm.traceID, tidLast)
-		}
-		isSeenTid = bm.traceID == tidLast
+	if bm.traceID < bw.tidLast {
+		logger.Panicf("the tid=%s cannot be smaller than the previously written tid=%s", bm.traceID, bw.tidLast)
 	}
-	bw.traceIDs = append(bw.traceIDs, bm.traceID)
+	hasWrittenBlocks := bw.hasWrittenBlocks
+	if !hasWrittenBlocks {
+		bw.tidFirst = bm.traceID
+		bw.hasWrittenBlocks = true
+	}
+	isSeenTid := bm.traceID == bw.tidLast
+	bw.tidLast = bm.traceID
+	if bw.traceIDFilter != nil && bw.traceIDFilter.filter != nil {
+		bw.traceIDFilter.filter.Add(convert.StringToBytes(bm.traceID))
+	}
 	bw.tagType.copyFrom(bm.tagType)
 
 	tm := &bm.timestamps
@@ -361,15 +388,9 @@ func (bw *blockWriter) Flush(pm *partMetadata, tf *traceIDFilter, tt *tagType) {
 
 	pm.CompressedSizeBytes = bw.writers.totalBytesWritten()
 
-	if len(bw.traceIDs) > 0 {
-		if tf.filter == nil {
-			tf.filter = generateTraceIDBloomFilter()
-		}
-		tf.filter.SetN(len(bw.traceIDs))
-		tf.filter.ResizeBits((len(bw.traceIDs)*filter.B + 63) / 64)
-		for _, traceID := range bw.traceIDs {
-			tf.filter.Add(convert.StringToBytes(traceID))
-		}
+	if bw.traceIDFilter != nil && bw.traceIDFilter.filter != nil {
+		tf.filter = bw.traceIDFilter.filter
+		bw.traceIDFilter.filter = nil
 	}
 	tt.copyFrom(bw.tagType)
 
@@ -380,7 +401,7 @@ func (bw *blockWriter) Flush(pm *partMetadata, tf *traceIDFilter, tt *tagType) {
 func generateBlockWriter() *blockWriter {
 	v := blockWriterPool.Get()
 	if v == nil {
-		return &blockWriter{
+		v = &blockWriter{
 			writers: writers{
 				tagMetadataWriters: make(map[string]*writer),
 				tagWriters:         make(map[string]*writer),